fix: move constant var about sql to sqls.go
